% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getQuery}
\alias{getQuery}
\title{Formats an RPostgreSQL query and then submits the query}
\usage{
getQuery(con, query)
}
\arguments{
\item{con}{A database connection object for RPostgreSQL.}

\item{query}{The query to be formatted.}
}
\description{
This function allows the text of a SQL query to be formatted across multiple
lines in an R document, by stripping the line breaks prior to submitting the
query based on provided \code{con} object.
}
