% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdr_create_json_catchment.R
\name{sdr_create_json_catchment}
\alias{sdr_create_json_catchment}
\title{Creates a GeoJSON probabilistic catchment for a station}
\usage{
sdr_create_json_catchment(crs, tablesuffix)
}
\arguments{
\item{crs}{crscode of the station}

\item{tablesuffix}{The suffix of the probability table - either crscode
(isolation) or 'concurrent' (concurrent) is expected.}
}
\description{
Creates a GeoJSON probabilistic catchment for a station. Includes a spatial
query to check that the postcode is within the 60 minute service area of the
station (important for "concurrent" mode). Only includes postcodes where the
probability of the station being chosen is > 1\%.
}
