% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{sdr_dbGetQuery}
\alias{sdr_dbGetQuery}
\title{Formats an RPostgres getQuery and then submits the query}
\usage{
sdr_dbGetQuery(con, query)
}
\arguments{
\item{con}{A database connection object for RPostgres.}

\item{query}{The query to be formatted.}
}
\description{
This function allows the text of a SQL query to be formatted across multiple
lines in an R document, by stripping the line breaks prior to submitting the
query based on provided \code{con} object.
}
