% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdr_frequency_group_adjustment.R
\name{sdr_frequency_group_adjustment}
\alias{sdr_frequency_group_adjustment}
\title{Adjusts the service frequency of stations based on frequency group}
\usage{
sdr_frequency_group_adjustment(schema, df, tablesuffix)
}
\arguments{
\item{schema}{A text string for the database schema name.}

\item{df}{A data frame containing the frequency group input. In isolation mode
this will relate to a specific station. In concurrent mode there will only be
a single frequency group across all stations.}

\item{tablesuffix}{The suffix of the probability table to be updated - either a crscode
(isolation) or 'concurrent' (concurrent) is expected.}
}
\description{
Updates the probability table for a station (isolation) or
stations (concurrent), adjusting the frequency variable for existing stations
based on information provided in the frequency group input file.
}
