% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdr_generate_choicesets_parallel.R
\name{sdr_generate_choicesets_parallel}
\alias{sdr_generate_choicesets_parallel}
\title{Generates a choice set of ten nearest stations for postcodes}
\usage{
sdr_generate_choicesets_parallel(crs)
}
\arguments{
\item{crs}{A character vector of the crscode of the station (isolation)
or stations (concurrent) as they appear in the model.proposed_stations table.}
}
\value{
Returns a data frame containing the postcode choicesets with stations ranked
by distance.
}
\description{
For a given proposed station (isolation) or stations (concurrent)
identifies the postcodes within 60 minutes of the station or stations and
then generates a choice set of the nearest 10 stations to each of those
postcodes.
}
\details{
The function relies on several pgRouting wrapper functions that must be
located in the openroads schema: \code{create_pgr_vnodes},
\code{sdr_crs_pc_nearest_stationswithpoints}, \code{sdr_pc_station_withpoints}, and
\code{sdr_pc_station_withpoints_nobbox}.

The function uses parallel processing via the foreach package and requires
the clusters to be configured prior to calling the function.
}
