% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdr_populate_probability_table.R
\name{sdr_populate_probability_table}
\alias{sdr_populate_probability_table}
\title{Populates the probability table of a station}
\usage{
sdr_populate_probability_table(tablesuffix)
}
\arguments{
\item{tablesuffix}{The suffix of the probability table - either crscode
(isolation) or 'concurrent' (concurrent) is expected.}
}
\description{
Populates the probability table of a station (isolation)
or stations (concurrent) with the variables required to apply
the station choice model. The variables are drawn from the model.proposed_stations
and data.stations tables.
}
