% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdr_create_service_areas.R
\name{sdr_create_service_areas}
\alias{sdr_create_service_areas}
\title{Creates a distance or time based network service area around stations}
\usage{
sdr_create_service_areas(df, sa, schema, table, cost = "len",
  target = 0.9)
}
\arguments{
\item{df}{A dataframe with a column called "location" which has easting and
northing of the station location.}

\item{sa}{A vector of integer values for the required service areas. Should
be in metres when cost is distance and minutes when cost is time.}

\item{schema}{A text string for the database schema name.}

\item{table}{A text string for the database table name.}

\item{cost}{A text string, either "len" for a distance-based service area or
"time" for a time-based service area. Default is "len".}

\item{target}{The target percent of area of convex hull that ST_ConvexHull
will try to approach before giving up or exiting. Default is 0.9.}
}
\description{
Creates a distance or time based network service area around stations in
the provided data frame using the pgRouting function \code{pgr_withpointsdd}. The
service area geometry is written to the specified database table in the
specified database schema.
}
