% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdr_generate_probability_table.R
\name{sdr_generate_probability_table}
\alias{sdr_generate_probability_table}
\title{Creates and populates the probability table for a station}
\usage{
sdr_generate_probability_table(con, schema, df, tablesuffix)
}
\arguments{
\item{con}{An RPostgres database connection object.}

\item{schema}{Character. for the database schema name.}

\item{df}{A dataframe containing the choicesets.}

\item{tablesuffix}{Character, the suffix of the probability table - either
crscode (isolation) or 'concurrent' (concurrent) is expected.}
}
\description{
Creates and populates the probability table for a station (isolation)
or stations (concurrent) with the variables required to apply
the station choice model. The variables are drawn from the
schema.proposed_stations and data.stations tables.
}
